Replace debounce with throttle in VolumeReceiver
The `VolumeReceiver` previously used a debounce mechanism to handle volume changes. This has been replaced with a throttle mechanism.

The constant `DEBOUNCE_TIME_MS` has been renamed to `THROTTLE_TIME_MS` and its value changed from `500` to `100L`.

The `VolumeReceiver`'s `onReceive` method now checks if the time since the last update is greater than or equal to `THROTTLE_TIME_MS`. If it is, `updateNotification` is called immediately. Otherwise, the event is ignored. The `Handler` and delayed `postDelayed` call have been removed.